using specific type for server port

+++ include/ws.h
@@ -18,6 +18,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>
 #define WS_H
 
 #include <stdbool.h>
+#include <stdint.h>
 
 #define MESSAGE_LENGTH 2048
 #define MAX_CLIENTS    8
@@ -63,6 +64,6 @@ extern int getHSresponse(char *hsrequest, char **hsresponse);
 
 extern char* ws_getaddress(int fd);
 extern int   ws_sendframe(int fd, const char *msg, bool broadcast);
-extern int   ws_socket(struct ws_events *evs, int port);
+extern int   ws_socket(struct ws_events *evs, uint16_t port);
 
 #endif /* WS_H */
